Mark a column of a table in a dm object as its primary key

cdm_add_pk() marks the given column as the given table's primary key in the data_model-part of the dm object. If check == TRUE, it also first checks if the given column is a unique key of the table. If force == TRUE, it replaces an already set key.

cdm_add_pk(dm, table, column, check = TRUE, force = FALSE)

Arguments

dm
A dm object.

table
A table in the dm

column
A column of that table

check
Boolean, if TRUE (default), a check is made if the column is a unique key of the table.

force
Boolean, if FALSE (default), an error will be thrown, if there is already a primary key set for this table. If TRUE a potential old pk is deleted before setting the new one.

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

nycflights_dm <- dm(src_df(pkg = "nycflights13"))

# the following works
cdm_add_pk(nycflights_dm, planes, tailnum)

#> ── Table source ─────────────────────────────────────────────────────────────────
#> src:  <package: nycflights13>
#> ── Data model ───────────────────────────────────────────────────────────────────
#> Data model object:
#>   5 tables:  airlines, airports, flights, planes ... 
#>   53 columns
#>   1 primary keys
#>   no references
#> ── Rows ─────────────────────────────────────────────────────────────────────────
#> Total: 367687
#> airlines: 16, airports: 1458, flights: 336776, planes: 3322, weather: 26115

cdm_add_pk(nycflights_dm, airports, faa)

#> ── Table source ─────────────────────────────────────────────────────────────────
#> src:  <package: nycflights13>
#> ── Data model ───────────────────────────────────────────────────────────────────
#> Data model object:
#>   5 tables:  airlines, airports, flights, planes ... 
#>   53 columns
#>   1 primary keys
#>   no references
#> ── Rows ─────────────────────────────────────────────────────────────────────────
#> Total: 367687
#> airlines: 16, airports: 1458, flights: 336776, planes: 3322, weather: 26115

cdm_add_pk(nycflights_dm, planes, manufacturer, check = FALSE)

#> ── Table source ─────────────────────────────────────────────────────────────────
#> src:  <package: nycflights13>
#> ── Data model ───────────────────────────────────────────────────────────────────
#> Data model object:
#>   5 tables:  airlines, airports, flights, planes ... 
#>   53 columns
#>   1 primary keys
#>   no references
#> ── Rows ─────────────────────────────────────────────────────────────────────────
#> Total: 367687
#> airlines: 16, airports: 1458, flights: 336776, planes: 3322, weather: 26115

# the following does not work
try(cdm_add_pk(nycflights_dm, planes, manufacturer))

#> Error : `manufacturer` not a unique key of `table_from_dm`.
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               ├─pkgdown:::as_data(...)
#>  20.               └─pkgdown:::as_data.tag_examples(...)
#>  21.                 └─purrr::pmap_chr(...)
#>  22.                   └─pkgdown:::.f(...)
#>  23.                     ├─withr::with_options(...)
#>  24.                     │ └─base::force(code)
#>  25.                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>  26.                       └─evaluate:::evaluate_call(...)
#>  27.                         ├─evaluate:::timing_fn(...)
#>  28.                         ├─evaluate:::handle(...)
#>  29.                         │ └─base::try(f, silent = TRUE)
#>  30.                         │   └─base::tryCatch(...)
#>  31.                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  32.                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  33.                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  34.                         ├─base::withCallingHandlers(...)
#>  35.                         ├─base::withVisible(eval(expr, envir, enclos))
#>  36.                         └─base::eval(expr, envir, enclos)
#>  37.                           └─base::eval(expr, envir, enclos)
#>  38.                             ├─base::try(cdm_add_pk(nycflights_dm, planes, manufacturer))
#>  39.                             │ └─base::tryCatch(...)
#>  40.                             │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  41.                             │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  42.                             │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  43.                             └─dm::cdm_add_pk(nycflights_dm, planes, manufacturer)
#>  44.                               └─dm::check_key(table_from_dm, !!col_expr) /home/kirill/git/cynkra/cynkra/public/dm/R/primary-keys.R:59:4
#>  45.                                 └─dm:::abort_not_unique_key(as_label(data_q), map_chr(args, as_label)) /home/kirill/git/cynkra/cynkra/public/dm/R/key-helpers.R:39:2