Find foreign key candidates in a table

Which columns are foreign candidates of a table, referencing the primary key column of another dm object's table? cdm_check_for_fk_candidates() checks first, if ref_table has a primary key set. Then it determines for each column of table, if this column contains only a subset of values of the primary key column of ref_table and is therefore a candidate for a foreign key from table to ref_table.

cdm_check_for_fk_candidates(dm, table, ref_table)

Arguments

dm
A dm object.

table
The table whose columns should be tested for foreign key candidate potential

ref_table
A table with a primary key.

Examples

library(dplyr)

nycflights_dm <- cdm_nycflights13()

nycflights_dm %>%
  cdm_check_for_fk_candidates(flights, airports)

#> # A tibble: 19 x 5
#>    candidate column         table   ref_table ref_table_pk
#>    <lgl>     <chr>          <chr>   <chr>     <chr>       
#>  1 FALSE     year           flights airports  faa         
#>  2 FALSE     month          flights airports  faa         
#>  3 FALSE     day            flights airports  faa         
#>  4 FALSE     dep_time       flights airports  faa         
#>  5 FALSE     sched_dep_time flights airports  faa         
#>  6 FALSE     dep_delay      flights airports  faa         
#>  7 FALSE     arr_time       flights airports  faa         
#>  8 FALSE     sched_arr_time flights airports  faa         
#>  9 FALSE     arr_delay      flights airports  faa         
#> 10 FALSE     carrier        flights airports  faa         
#> 11 FALSE     flight         flights airports  faa         
#> 12 FALSE     tailnum        flights airports  faa         
#> 13 TRUE      origin         flights airports  faa         
#> 14 FALSE     dest           flights airports  faa         
#> 15 FALSE     air_time       flights airports  faa         
#> 16 FALSE     distance       flights airports  faa         
#> 17 FALSE     hour           flights airports  faa         
#> 18 FALSE     minute         flights airports  faa         
#> 19 FALSE     time_hour      flights airports  faa