Find foreign key candidates in a table

Determine which columns would be good candidates to be used as foreign keys of a table, and reference the primary key column of another dm object's table.

cdm_enum_fk_candidates(dm, table, ref_table)

enum_fk_candidates(zoomed_dm, ref_table)

Arguments

dm
A dm object.

table
The table whose columns should be tested for suitability of use as foreign keys.

ref_table
A table with a primary key.

zoomed_dm
A dm with a zoomed table.

Value

A table that lists which columns of table would be suitable candidates for foreign key columns to reference ref_table, and which columns would not be suitable.

Details

cdm_enum_fk_candidates() first checks if ref_table has a primary key set. Then, for each column of table, a join operation will be tried, with argument by matching the respective column with the primary key of ref_table. This tests implicitly for type compatibility (on most sources). Based on the result of the join, the entry in the result column why is:

an empty entry, if the column is a candidate

the total percentage and individual numbers of missing matches between the entries of the respective column in table table and the primary column entries in table ref_table.

the error message triggered by the error (often stating the mismatched column types)

enum_fk_candidates() works like cdm_enum_fk_candidates() with the zoomed table as table.

See also

Examples

cdm_enum_fk_candidates(cdm_nycflights13(), flights, airports)

#> # A tibble: 19 x 3
#>    column       candidate why                                                   
#>    <chr>        <lgl>     <chr>                                                 
#>  1 origin       TRUE      ""                                                    
#>  2 dest         FALSE     7602 entries (2.3%) of `flights$dest` not in `airport…
#>  3 tailnum      FALSE     334264 entries (99.3%) of `flights$tailnum` not in `a…
#>  4 carrier      FALSE     336776 entries (100%) of `flights$carrier` not in `ai…
#>  5 air_time     FALSE     Can't join on 'value' x 'value' because of incompatib…
#>  6 arr_delay    FALSE     Can't join on 'value' x 'value' because of incompatib…
#>  7 arr_time     FALSE     Can't join on 'value' x 'value' because of incompatib…
#>  8 day          FALSE     Can't join on 'value' x 'value' because of incompatib…
#>  9 dep_delay    FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 10 dep_time     FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 11 distance     FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 12 flight       FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 13 hour         FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 14 minute       FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 15 month        FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 16 sched_arr_t… FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 17 sched_dep_t… FALSE     Can't join on 'value' x 'value' because of incompatib…
#> 18 time_hour    FALSE     cannot join a POSIXct object with an object that is n…
#> 19 year         FALSE     Can't join on 'value' x 'value' because of incompatib…