Flatten part of a dm into a wide table

This function joins all the tables of your dm object together, that can be reached from table start in the direction that the foreign keys are pointing, using the foreign key relations to determine the parameter by for the necessary joins. It returns one table with unique columns. Use cdm_select_tbl() if necessary to reduce the number of tables before calling this function.

cdm_flatten_to_tbl(dm, start, join = left_join)

Arguments

dm
A dm object

start
Table to start from. From this table all outgoing foreign key relations are considered to establish a processing order for the joins. An interesting choice could be for example a fact table in a star schema.

join
The type of join to be performed, see join

Value

A wide table resulting of consecutively joining all tables together.

Details

Uses cdm_apply_filters() and cdm_disambiguate_cols() first, to get a "clean" dm. Subsequently renames all foreign key columns to the names of the primary key columns they are pointing to. Then the order of the joins is determined and the joins are performed.

See also

Examples

cdm_nycflights13() %>%
cdm_select_tbl(-weather) %>%
cdm_flatten_to_tbl(flights)

#> Renamed columns:
#> Table: airlines
#> name -> airlines.name
#> 
#> Table: airports
#> name -> airports.name
#> 
#> Table: flights
#> year -> flights.year
#> 
#> Table: planes
#> year -> planes.year

#> # A tibble: 336,776 x 35
#>    flights.year month   day dep_time sched_dep_time dep_delay arr_time
#>           <int> <int> <int>    <int>          <int>     <dbl>    <int>
#>  1         2013     1     1      517            515         2      830
#>  2         2013     1     1      533            529         4      850
#>  3         2013     1     1      542            540         2      923
#>  4         2013     1     1      544            545        -1     1004
#>  5         2013     1     1      554            600        -6      812
#>  6         2013     1     1      554            558        -4      740
#>  7         2013     1     1      555            600        -5      913
#>  8         2013     1     1      557            600        -3      709
#>  9         2013     1     1      557            600        -3      838
#> 10         2013     1     1      558            600        -2      753
#> # … with 336,766 more rows, and 28 more variables: sched_arr_time <int>,
#> #   arr_delay <dbl>, carrier <chr>, flight <int>, tailnum <chr>, faa <chr>,
#> #   dest <chr>, air_time <dbl>, distance <dbl>, hour <dbl>, minute <dbl>,
#> #   time_hour <dttm>, airlines.name <chr>, airports.name <chr>, lat <dbl>,
#> #   lon <dbl>, alt <int>, tz <dbl>, dst <chr>, tzone <chr>, planes.year <int>,
#> #   type <chr>, manufacturer <chr>, model <chr>, engines <int>, seats <int>,
#> #   speed <int>, engine <chr>