Perform a join between two tables of a dm

A join of desired type is performed between table table_1 and table table_2. The two tables need to be directly connected by a foreign key relation. Since this function is a wrapper around cdm_flatten_to_tbl(), the LHS of the join will always be the "child table", the table referencing the other table.

cdm_join_to_tbl(dm, table_1, table_2, join = left_join)

Arguments

dm
A dm object

table_1
One of the tables involved in the join

table_2
The second table of the join

join
The type of join to be performed, see dplyr::join()

Value

The resulting table of the join.

See also