Data model class

The dm class wraps dplyr::src and adds a description of table relationships based on datamodelr::datamodelr-package.

dm() coerces its inputs.

new_dm() only checks if the inputs are of the correct class.

validate_dm() checks consistency between the dplyr source and the datamodelr based specification of table relationships.

cdm_get_src() returns the dplyr source component of a dm object.

cdm_get_tables() returns a named list with dplyr tbl objects of a dm object.

cdm_get_data_model() returns the datamodelr data model component of a dm object.

is_dm() returns TRUE if the input is of class dm.

as_dm() coerces objects to the dm class

dm(src, data_model = NULL)

new_dm(src, tables, data_model)

validate_dm(x)

cdm_get_src(x)

cdm_get_tables(x)

cdm_get_data_model(x)

is_dm(x)

as_dm(x)

Arguments

src
A dplyr table source object.

data_model
A datamodelr data model object, or NULL.

tables
A list of the tables (tibble-objects, not names) to be included in the dm object

x
An object.