dm: Relational Data Models

Provides tools for working with multiple related tables, stored as data frames or in a relational database. Multiple tables (data and metadata) are stored in a compound object, which can then be manipulated with a pipe-friendly syntax.

See also

Author

Maintainer: Kirill Müller kirill@cynkra.com (ORCID)

Authors:

Tobias Schieferdecker tobias@cynkra.com

Darko Bergant darko.bergant@gmail.com

Other contributors:

Antoine Fabri [contributor]

Katharina Brunner [contributor]

James Wondrasek [contributor]

Indrajeet Patil (ORCID) [contributor]

Maëlle Salmon (ORCID) [contributor]

energie360° AG info@energie360.ch [funder]

cynkra GmbH mail@cynkra.com (ROR) [funder, copyright holder]