Get the names of referencing tables

This function returns the names of all tables that point to the primary key of a table.

dm_get_referencing_tables(dm, table)

Arguments

dm
A dm object.

table
A table in the dm.

Value

A character vector of the names of the tables that point to the primary key of table.

See also

Examples

dm_get_referencing_tables(dm_nycflights13(), airports)

#> [1] "flights"

dm_get_referencing_tables(dm_nycflights13(), flights)

#> character(0)