Update tables in a dm

Updates one or more existing tables in a dm. For now, the column names must be identical. This restriction may be levied optionally in the future.

dm_mutate_tbl(dm, ...)

Arguments

dm

A dm object.

...

One or more tables to update in the dm. Must be named.

See also

Examples

dm_nycflights13() %>%
  dm_mutate_tbl(flights = nycflights13::flights[1:3, ])
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `airlines`, `airports`, `flights`, `planes`, `weather`
#> Columns: 53
#> Primary keys: 4
#> Foreign keys: 4