Creates a dm object for the pixarfilms data

Creates an example dm object from the tables in pixarfilms, along with the references.

Usage

dm_pixarfilms(..., color = TRUE, consistent = FALSE)

Arguments

...

These dots are for future extensions and must be empty.

color

Boolean, if TRUE (default), the resulting dm object will have colors assigned to different tables for visualization with dm_draw().

consistent

Boolean, In the original dm the film column in pixar_films contains missing values so cannot be made a proper primary key. Set to TRUE to remove those records.

Value

A dm object consisting of pixarfilms tables, complete with primary and foreign keys and optionally colored.

Examples

dm_pixarfilms()
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `pixar_films`, `pixar_people`, `academy`, `box_office`, `genres`, `public_response`
#> Columns: 23
#> Primary keys: 5
#> Foreign keys: 5
dm_pixarfilms() %>%
  dm_draw()