dplyr table retrieval, table info and DB interaction methods

Use these methods without the '.dm' or '.zoomed_dm' suffix (see examples).

# S3 method for dm
tbl(src, from, ...)

# S3 method for dm
compute(x, ...)

# S3 method for zoomed_dm
compute(x, ...)

# S3 method for dm
src_tbls(x)

# S3 method for dm
copy_to(
  dest,
  df,
  name = deparse(substitute(df)),
  overwrite = FALSE,
  temporary = TRUE,
  repair = "unique",
  quiet = FALSE,
  ...
)

# S3 method for dm
collect(x, ...)

Arguments

src
A dm object.

from
A length one character variable containing the name of the requested table

...
See original function documentation

x
Either a dm or a zoomed_dm; the latter leads to an error for src_tbls.dm()

dest
For copy_to.dm(): The dm object to which a table should be copied.

df
For copy_to.dm(): A table (can be on a different src)

name
For copy_to.dm(): See dplyr::copy_to

overwrite
For copy_to.dm(): See dplyr::copy_to; TRUE leads to an error

temporary
For copy_to.dm(): If the dm is on a DB, the copied version of df will only be written temporarily to the DB. After the connection is reset it will no longer be available.

repair, quiet
Name repair options; cf. vctrs::vec_as_names