dm as data source

These methods are deprecated because of their limited use, and because the notion of a "source" seems to be getting phased out from dplyr. Use other ways to access the tables in a dm.

dm_get_src(x)

# S3 method for dm
tbl(src, from, ...)

# S3 method for dm
src_tbls(x, ...)

# S3 method for dm
copy_to(
  dest,
  df,
  name = deparse(substitute(df)),
  overwrite = FALSE,
  temporary = TRUE,
  repair = "unique",
  quiet = FALSE,
  ...
)

Arguments

src

A dm object.

from

A length one character variable containing the name of the requested table

...

See original function documentation

dest

For copy_to.dm(): The dm object to which a table should be copied.

df

For copy_to.dm(): A table (can be on a different src)

name

For copy_to.dm(): See dplyr::copy_to

overwrite

For copy_to.dm(): See dplyr::copy_to; TRUE leads to an error

temporary

For copy_to.dm(): If the dm is on a DB, the copied version of df will only be written temporarily to the DB. After the connection is reset it will no longer be available.

repair, quiet

Name repair options; cf. vctrs::vec_as_names()

Details

Use dm_get_con() instead of dm_get_src() to get the DBI connetion for a dm object

Use [[ instead of tbl() to access individual tables in a dm object.

Get the names from dm_get_tables() instead of calling dm_get_src() to list the table names in a dm object.

Use copy_to() on a table and then dm_add_tbl() instead of copy_to() on a dm object.