Which columns are candidates for a primary key column?

enum_pk_candidates() checks for each column of a table if this column contains only unique values and is therefore a candidate for a primary key of this table.

cdm_enum_pk_candidates() performs these checks for a table in a dm object.

enum_pk_candidates(table)

cdm_enum_pk_candidates(dm, table)

Arguments

table
A table in the dm

dm
A dm object.

See also

Examples

nycflights13::flights %>% enum_pk_candidates()

#> # A tibble: 19 x 3
#>    column       candidate why                                                   
#>    <chr>        <lgl>     <chr>                                                 
#>  1 year         FALSE     has duplicate values: 2013                            
#>  2 month        FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  3 day          FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  4 dep_time     FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  5 sched_dep_t… FALSE     has duplicate values: 500, 505, 510, 515, 516, 517, … 
#>  6 dep_delay    FALSE     has duplicate values: -25, -24, -23, -22, -21, -20, … 
#>  7 arr_time     FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  8 sched_arr_t… FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  9 arr_delay    FALSE     has duplicate values: -75, -71, -70, -69, -68, -67, … 
#> 10 carrier      FALSE     has duplicate values: 9E, AA, AS, B6, DL, EV, …       
#> 11 flight       FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#> 12 tailnum      FALSE     has duplicate values: D942DN, N0EGMQ, N10156, N102UW,…
#> 13 origin       FALSE     has duplicate values: EWR, JFK, LGA                   
#> 14 dest         FALSE     has duplicate values: ABQ, ACK, ALB, ANC, ATL, AUS, … 
#> 15 air_time     FALSE     has duplicate values: 20, 21, 22, 23, 24, 25, …       
#> 16 distance     FALSE     has duplicate values: 80, 94, 96, 116, 143, 160, …    
#> 17 hour         FALSE     has duplicate values: 5, 6, 7, 8, 9, 10, …            
#> 18 minute       FALSE     has duplicate values: 0, 1, 2, 3, 4, 5, …             
#> 19 time_hour    FALSE     has duplicate values: 2013-01-01 05:00:00, 2013-01-01…

cdm_nycflights13() %>% cdm_enum_pk_candidates(flights)

#> # A tibble: 19 x 3
#>    column       candidate why                                                   
#>    <chr>        <lgl>     <chr>                                                 
#>  1 year         FALSE     has duplicate values: 2013                            
#>  2 month        FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  3 day          FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  4 dep_time     FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  5 sched_dep_t… FALSE     has duplicate values: 500, 505, 510, 515, 516, 517, … 
#>  6 dep_delay    FALSE     has duplicate values: -25, -24, -23, -22, -21, -20, … 
#>  7 arr_time     FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  8 sched_arr_t… FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#>  9 arr_delay    FALSE     has duplicate values: -75, -71, -70, -69, -68, -67, … 
#> 10 carrier      FALSE     has duplicate values: 9E, AA, AS, B6, DL, EV, …       
#> 11 flight       FALSE     has duplicate values: 1, 2, 3, 4, 5, 6, …             
#> 12 tailnum      FALSE     has duplicate values: D942DN, N0EGMQ, N10156, N102UW,…
#> 13 origin       FALSE     has duplicate values: EWR, JFK, LGA                   
#> 14 dest         FALSE     has duplicate values: ABQ, ACK, ALB, ANC, ATL, AUS, … 
#> 15 air_time     FALSE     has duplicate values: 20, 21, 22, 23, 24, 25, …       
#> 16 distance     FALSE     has duplicate values: 80, 94, 96, 116, 143, 160, …    
#> 17 hour         FALSE     has duplicate values: 5, 6, 7, 8, 9, 10, …            
#> 18 minute       FALSE     has duplicate values: 0, 1, 2, 3, 4, 5, …             
#> 19 time_hour    FALSE     has duplicate values: 2013-01-01 05:00:00, 2013-01-01…

cdm_nycflights13() %>% cdm_enum_pk_candidates(airports)

#> # A tibble: 8 x 3
#>   column candidate why                                                          
#>   <chr>  <lgl>     <chr>                                                        
#> 1 faa    TRUE      ""                                                           
#> 2 name   FALSE     has duplicate values: All Airports             , Capital Cit…
#> 3 lat    FALSE     has duplicate values: 38.88944, 40.63975                     
#> 4 lon    TRUE      ""                                                           
#> 5 alt    FALSE     has duplicate values: 0, 1, 3, 4, 5, 6, …                    
#> 6 tz     FALSE     has duplicate values: -10, -9, -8, -7, -6, -5, …             
#> 7 dst    FALSE     has duplicate values: A, N, U                                
#> 8 tzone  FALSE     "has duplicate values: \\N                , America/Anchorag…