Materialize

compute() materializes all tables in a dm to new (temporary or permanent) tables on the database.

collect() downloads the tables in a dm object as local tibbles.

# S3 method for dm
compute(x, ...)

# S3 method for dm
collect(x, ...)

Arguments

x
A dm.

...
Passed on to compute().

Value

A dm object of the same structure as the input.

Details

Called on a dm object, these methods create a copy of all tables in the dm. Depending on the size of your data this may take a long time.

Examples

financial <- dm_financial_sqlite()

financial %>%
  pull_tbl(districts) %>%
  dbplyr::remote_name()
#> <IDENT> `districts`

# compute() copies the data to new tables:
financial %>%
  compute() %>%
  pull_tbl(districts) %>%
  dbplyr::remote_name()
#> NULL

# collect() returns a local dm:
financial %>%
  collect() %>%
  pull_tbl(districts) %>%
  class()
#> [1] "tbl_df"     "tbl"        "data.frame"