utils table manipulation methods for zoomed_dm objects

Extract the first or last rows from a table. Use these methods without the '.zoomed_dm' suffix (see examples). The methods for regular dm objects extract the first or last tables.

# S3 method for zoomed_dm
head(x, n = 6L, ...)

# S3 method for zoomed_dm
tail(x, n = 6L, ...)

Arguments

x
object of class zoomed_dm

n
a single integer. If positive or zero, size for the resulting object: number of elements for a vector (including lists), rows for a matrix or data frame or lines for a function. If negative, all but the n last/first number of elements of x.

...
arguments to be passed to or from other methods.

Value

A zoomed_dm object.

Details

see manual for the corresponding functions in utils.

Examples

zoomed <- dm_nycflights13() %>%
  dm_zoom_to(flights) %>%
  head(4)
zoomed

#> # Zoomed table: flights
#> # A tibble:     4 x 19
#>    year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>   <int> <int> <int>    <int>          <int>     <dbl>    <int>          <int>
#> 1  2013     1    10        3           2359         4      426            437
#> 2  2013     1    10       16           2359        17      447            444
#> 3  2013     1    10      450            500       -10      634            648
#> 4  2013     1    10      520            525        -5      813            820
#> # … with 11 more variables: arr_delay <dbl>, carrier <chr>, flight <int>,
#> #   tailnum <chr>, origin <chr>, dest <chr>, air_time <dbl>, distance <dbl>,
#> #   hour <dbl>, minute <dbl>, time_hour <dttm>

dm_insert_zoomed(zoomed, new_tbl_name = "head_flights")

#> ── Table source ────────────────────────────────────────────────────────────────
#> src:  <environment: R_GlobalEnv>
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `airlines`, `airports`, `flights`, `planes`, `weather`, `head_flights`
#> Columns: 72
#> Primary keys: 3
#> Foreign keys: 6