Find foreign key candidates in a table

Which columns are foreign candidates of a table, referencing the primary key column of another dm object's table? cdm_enum_fk_candidates() checks first, if ref_table has a primary key set. Then it determines for each column of table, if this column contains only a subset of values of the primary key column of ref_table and is therefore a candidate for a foreign key from table to ref_table.

cdm_enum_fk_candidates(dm, table, ref_table)

Arguments

dm
A dm object.

table
The table whose columns should be tested for foreign key candidate potential

ref_table
A table with a primary key.

See also

Examples

library(dplyr)

nycflights_dm <- cdm_nycflights13(cycle = TRUE)

nycflights_dm %>%
  cdm_enum_fk_candidates(flights, airports)

#> # A tibble: 19 x 6
#>    ref_table ref_table_pk table   column       candidate why                    
#>    <chr>     <chr>        <chr>   <chr>        <lgl>     <chr>                  
#>  1 airports  faa          flights year         FALSE     not a subset of airpor…
#>  2 airports  faa          flights month        FALSE     not a subset of airpor…
#>  3 airports  faa          flights day          FALSE     not a subset of airpor…
#>  4 airports  faa          flights dep_time     FALSE     not a subset of airpor…
#>  5 airports  faa          flights sched_dep_t… FALSE     not a subset of airpor…
#>  6 airports  faa          flights dep_delay    FALSE     not a subset of airpor…
#>  7 airports  faa          flights arr_time     FALSE     not a subset of airpor…
#>  8 airports  faa          flights sched_arr_t… FALSE     not a subset of airpor…
#>  9 airports  faa          flights arr_delay    FALSE     not a subset of airpor…
#> 10 airports  faa          flights carrier      FALSE     not a subset of airpor…
#> 11 airports  faa          flights flight       FALSE     not a subset of airpor…
#> 12 airports  faa          flights tailnum      FALSE     not a subset of airpor…
#> 13 airports  faa          flights origin       TRUE      ""                     
#> 14 airports  faa          flights dest         FALSE     not a subset of airpor…
#> 15 airports  faa          flights air_time     FALSE     not a subset of airpor…
#> 16 airports  faa          flights distance     FALSE     not a subset of airpor…
#> 17 airports  faa          flights hour         FALSE     not a subset of airpor…
#> 18 airports  faa          flights minute       FALSE     not a subset of airpor…
#> 19 airports  faa          flights time_hour    FALSE     not a subset of airpor…