Cascading filter of a dm object

'cdm_filter()' allows you to set one or more filter conditions for one table of a dm object. If the dm's tables are connected via key constrains, the filtering will affect all other connected tables, leaving only the rows with the corresponding key values.

'cdm_semi_join()' performs a cascading "row reduction" of a dm object by an inital semi-join of one of its tables with the same, but filtered table. Subsequently, the key constraints are used to compute the remainders of the other tables of the dm object and a new dm object is returned.

cdm_filter(dm, table, ...)

cdm_semi_join(dm, table, reduced_table)

Arguments

dm
A dm object.

table
A table in the dm

...
Logical predicates defined in terms of the variables in .data. Multiple conditions are combined with &. Only rows where the condition evaluates to TRUE are kept.

The arguments in ... are automatically quoted and evaluated in the context of the data frame. They support unquoting and splicing. See vignette("programming") for an introduction to these concepts.

reduced_table
The table indicated in argument table, but in a filtered state (cf, dplyr::filter()).

Examples

library(magrittr)

#> 
#> Attaching package: ‘magrittr’

#> The following objects are masked from ‘package:testthat’:
#> 
#>     equals, is_less_than, not

cdm_nycflights13(cycle = FALSE) %>%
  cdm_filter(airports, name == "John F Kennedy Intl")

#> ── Table source ────────────────────────────────────────────────────────────────
#> src:  <package: nycflights13>
#> ── Data model ──────────────────────────────────────────────────────────────────
#> Data model object:
#>   5 tables:  airlines, airports, flights, planes ... 
#>   53 columns
#>   1 primary keys
#>   3 references
#> ── Rows ────────────────────────────────────────────────────────────────────────
#> Total: 138786
#> airlines: 10, airports: 1, flights: 111279, planes: 1381, weather: 26115