Data model class

The dm class holds a list of tables and their relationships. It is inspired by datamodelr, and extends the idea by offering operations to access the data in the tables.

dm() creates a dm object from one or multiple tbl objects (tibbles or lazy data objects).

dm_from_src() creates a dm from some or all tables in a src (a database or an environment).

new_dm() is a low-level constructor that creates a new dm object.

If called without arguments, it will create an empty dm.

If called with arguments, no validation checks will be made to ascertain that the inputs are of the expected class and internally consistent; use validate_dm() to double-check the returned object.

validate_dm() checks the internal consistency of a dm object.

dm_get_src() returns the dplyr source for a dm object. All tables in a dm object must be from the same source, i.e. either they are all data frames, or they all are stored on the same database.

dm_get_con() returns the DBI::DBIConnection for dm objects. This works only if the tables are stored on a database, otherwise an error is thrown.

dm_get_tables() returns a named list of dplyr tbl objects of a dm object. Filtering expressions are NOT evaluated at this stage. To get a filtered table, use dm_apply_filters_to_tbl(), to apply filters to all tables use dm_apply_filters()

dm_get_filter() returns the filter expressions that have been applied to a dm object. These filter expressions are not intended for evaluation, only for information.

is_dm() returns TRUE if the input is of class dm.

as_dm() coerces objects to the dm class

dm(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dm_from_src(src = NULL, table_names = NULL)

new_dm(tables = list())

validate_dm(x)

dm_get_src(x)

dm_get_con(x)

dm_get_tables(x)

dm_get_filter(x)

is_dm(x)

as_dm(x)

Arguments

...
Tables to add to the dm object. If no names are provided, the tables are auto-named.

.name_repair
Options for name repair. Forwarded as repair to vctrs::vec_as_names().

src
A dplyr table source object.

table_names
A character vector of the names of the tables to include.

tables
A named list of the tables (tibble-objects, not names) . to be included in the dm object.

x
An object.

Value

For dm(), dm_from_src(), new_dm(), as_dm(): A dm object.

For validate_dm(): Returns the dm, invisibly, after finishing all checks.

For dm_get_src(): the dplyr source for a dm object.

For dm_get_con(): The DBI::DBIConnection for dm objects.

For dm_get_tables(): A named list with the tables constituing the dm.

For dm_get_filter(): A tibble with columns:

"table": table that was filtered,

"filter": the filter expression,

"zoomed": logical, does the filter condition relate to the zoomed table.

For is_dm(): Boolean, is this object a dm.

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

dm(iris, mtcars)

#> ── Table source ────────────────────────────────────────────────────────────────
#> src:  <environment: R_GlobalEnv>
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `iris`, `mtcars`
#> Columns: 16
#> Primary keys: 0
#> Foreign keys: 0

dm_from_src(dplyr::src_df(pkg = "nycflights13"))

#> ── Table source ────────────────────────────────────────────────────────────────
#> src:  <environment: R_GlobalEnv>
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `airlines`, `airports`, `flights`, `planes`, `weather`
#> Columns: 53
#> Primary keys: 0
#> Foreign keys: 0

new_dm(list(iris = iris, mtcars = mtcars))

#> ── Table source ────────────────────────────────────────────────────────────────
#> src:  <environment: R_GlobalEnv>
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `iris`, `mtcars`
#> Columns: 16
#> Primary keys: 0
#> Foreign keys: 0

as_dm(list(iris = iris, mtcars = mtcars))

#> ── Table source ────────────────────────────────────────────────────────────────
#> src:  <environment: R_GlobalEnv>
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `iris`, `mtcars`
#> Columns: 16
#> Primary keys: 0
#> Foreign keys: 0

dm_nycflights13() %>% tbl("airports")

#> # A tibble: 1,458 x 8
#>    faa   name                       lat    lon   alt    tz dst   tzone          
#>    <chr> <chr>                    <dbl>  <dbl> <dbl> <dbl> <chr> <chr>          
#>  1 04G   Lansdowne Airport         41.1  -80.6  1044    -5 A     America/New_Yo…
#>  2 06A   Moton Field Municipal A…  32.5  -85.7   264    -6 A     America/Chicago
#>  3 06C   Schaumburg Regional       42.0  -88.1   801    -6 A     America/Chicago
#>  4 06N   Randall Airport           41.4  -74.4   523    -5 A     America/New_Yo…
#>  5 09J   Jekyll Island Airport     31.1  -81.4    11    -5 A     America/New_Yo…
#>  6 0A9   Elizabethton Municipal …  36.4  -82.2  1593    -5 A     America/New_Yo…
#>  7 0G6   Williams County Airport   41.5  -84.5   730    -5 A     America/New_Yo…
#>  8 0G7   Finger Lakes Regional A…  42.9  -76.8   492    -5 A     America/New_Yo…
#>  9 0P2   Shoestring Aviation Air…  39.8  -76.6  1000    -5 U     America/New_Yo…
#> 10 0S9   Jefferson County Intl     48.1 -123.    108    -8 A     America/Los_An…
#> # … with 1,448 more rows

dm_nycflights13() %>% src_tbls()

#> [1] "airlines" "airports" "flights"  "planes"   "weather" 

dm_nycflights13() %>% dm_get_src()

#> src:  <environment: R_GlobalEnv>
#> tbls:

# this works only when tables of `dm` are on DB
if (FALSE) {
  dm_copy_to(dbplyr::src_memdb(), dm_nycflights13()) %>%
    dm_get_con()
}
dm_nycflights13() %>% dm_get_tables()

#> $airlines
#> # A tibble: 16 x 2
#>    carrier name                       
#>    <chr>   <chr>                      
#>  1 9E      Endeavor Air Inc.          
#>  2 AA      American Airlines Inc.     
#>  3 AS      Alaska Airlines Inc.       
#>  4 B6      JetBlue Airways            
#>  5 DL      Delta Air Lines Inc.       
#>  6 EV      ExpressJet Airlines Inc.   
#>  7 F9      Frontier Airlines Inc.     
#>  8 FL      AirTran Airways Corporation
#>  9 HA      Hawaiian Airlines Inc.     
#> 10 MQ      Envoy Air                  
#> 11 OO      SkyWest Airlines Inc.      
#> 12 UA      United Air Lines Inc.      
#> 13 US      US Airways Inc.            
#> 14 VX      Virgin America             
#> 15 WN      Southwest Airlines Co.     
#> 16 YV      Mesa Airlines Inc.         
#> 
#> $airports
#> # A tibble: 1,458 x 8
#>    faa   name                       lat    lon   alt    tz dst   tzone          
#>    <chr> <chr>                    <dbl>  <dbl> <dbl> <dbl> <chr> <chr>          
#>  1 04G   Lansdowne Airport         41.1  -80.6  1044    -5 A     America/New_Yo…
#>  2 06A   Moton Field Municipal A…  32.5  -85.7   264    -6 A     America/Chicago
#>  3 06C   Schaumburg Regional       42.0  -88.1   801    -6 A     America/Chicago
#>  4 06N   Randall Airport           41.4  -74.4   523    -5 A     America/New_Yo…
#>  5 09J   Jekyll Island Airport     31.1  -81.4    11    -5 A     America/New_Yo…
#>  6 0A9   Elizabethton Municipal …  36.4  -82.2  1593    -5 A     America/New_Yo…
#>  7 0G6   Williams County Airport   41.5  -84.5   730    -5 A     America/New_Yo…
#>  8 0G7   Finger Lakes Regional A…  42.9  -76.8   492    -5 A     America/New_Yo…
#>  9 0P2   Shoestring Aviation Air…  39.8  -76.6  1000    -5 U     America/New_Yo…
#> 10 0S9   Jefferson County Intl     48.1 -123.    108    -8 A     America/Los_An…
#> # … with 1,448 more rows
#> 
#> $flights
#> # A tibble: 336,776 x 19
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <int> <int> <int>    <int>          <int>     <dbl>    <int>          <int>
#>  1  2013     1     1      517            515         2      830            819
#>  2  2013     1     1      533            529         4      850            830
#>  3  2013     1     1      542            540         2      923            850
#>  4  2013     1     1      544            545        -1     1004           1022
#>  5  2013     1     1      554            600        -6      812            837
#>  6  2013     1     1      554            558        -4      740            728
#>  7  2013     1     1      555            600        -5      913            854
#>  8  2013     1     1      557            600        -3      709            723
#>  9  2013     1     1      557            600        -3      838            846
#> 10  2013     1     1      558            600        -2      753            745
#> # … with 336,766 more rows, and 11 more variables: arr_delay <dbl>,
#> #   carrier <chr>, flight <int>, tailnum <chr>, origin <chr>, dest <chr>,
#> #   air_time <dbl>, distance <dbl>, hour <dbl>, minute <dbl>, time_hour <dttm>
#> 
#> $planes
#> # A tibble: 3,322 x 9
#>    tailnum  year type          manufacturer   model  engines seats speed engine 
#>    <chr>   <int> <chr>         <chr>          <chr>    <int> <int> <int> <chr>  
#>  1 N10156   2004 Fixed wing m… EMBRAER        EMB-1…       2    55    NA Turbo-…
#>  2 N102UW   1998 Fixed wing m… AIRBUS INDUST… A320-…       2   182    NA Turbo-…
#>  3 N103US   1999 Fixed wing m… AIRBUS INDUST… A320-…       2   182    NA Turbo-…
#>  4 N104UW   1999 Fixed wing m… AIRBUS INDUST… A320-…       2   182    NA Turbo-…
#>  5 N10575   2002 Fixed wing m… EMBRAER        EMB-1…       2    55    NA Turbo-…
#>  6 N105UW   1999 Fixed wing m… AIRBUS INDUST… A320-…       2   182    NA Turbo-…
#>  7 N107US   1999 Fixed wing m… AIRBUS INDUST… A320-…       2   182    NA Turbo-…
#>  8 N108UW   1999 Fixed wing m… AIRBUS INDUST… A320-…       2   182    NA Turbo-…
#>  9 N109UW   1999 Fixed wing m… AIRBUS INDUST… A320-…       2   182    NA Turbo-…
#> 10 N110UW   1999 Fixed wing m… AIRBUS INDUST… A320-…       2   182    NA Turbo-…
#> # … with 3,312 more rows
#> 
#> $weather
#> # A tibble: 26,115 x 15
#>    origin  year month   day  hour  temp  dewp humid wind_dir wind_speed
#>    <chr>  <int> <int> <int> <int> <dbl> <dbl> <dbl>    <dbl>      <dbl>
#>  1 EWR     2013     1     1     1  39.0  26.1  59.4      270      10.4 
#>  2 EWR     2013     1     1     2  39.0  27.0  61.6      250       8.06
#>  3 EWR     2013     1     1     3  39.0  28.0  64.4      240      11.5 
#>  4 EWR     2013     1     1     4  39.9  28.0  62.2      250      12.7 
#>  5 EWR     2013     1     1     5  39.0  28.0  64.4      260      12.7 
#>  6 EWR     2013     1     1     6  37.9  28.0  67.2      240      11.5 
#>  7 EWR     2013     1     1     7  39.0  28.0  64.4      240      15.0 
#>  8 EWR     2013     1     1     8  39.9  28.0  62.2      250      10.4 
#>  9 EWR     2013     1     1     9  39.9  28.0  62.2      260      15.0 
#> 10 EWR     2013     1     1    10  41    28.0  59.6      260      13.8 
#> # … with 26,105 more rows, and 5 more variables: wind_gust <dbl>, precip <dbl>,
#> #   pressure <dbl>, visib <dbl>, time_hour <dttm>
#> 

dm_nycflights13() %>% dm_get_filter()

#> # A tibble: 0 x 3
#> # … with 3 variables: table <chr>, filter <list>, zoomed <lgl>

dm_nycflights13() %>% validate_dm()
is_dm(dm_nycflights13())

#> [1] TRUE

dm_nycflights13()["airports"]

#> ── Table source ────────────────────────────────────────────────────────────────
#> src:  <environment: R_GlobalEnv>
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `airports`
#> Columns: 8
#> Primary keys: 1
#> Foreign keys: 0

dm_nycflights13()[["airports"]]

#> # A tibble: 1,458 x 8
#>    faa   name                       lat    lon   alt    tz dst   tzone          
#>    <chr> <chr>                    <dbl>  <dbl> <dbl> <dbl> <chr> <chr>          
#>  1 04G   Lansdowne Airport         41.1  -80.6  1044    -5 A     America/New_Yo…
#>  2 06A   Moton Field Municipal A…  32.5  -85.7   264    -6 A     America/Chicago
#>  3 06C   Schaumburg Regional       42.0  -88.1   801    -6 A     America/Chicago
#>  4 06N   Randall Airport           41.4  -74.4   523    -5 A     America/New_Yo…
#>  5 09J   Jekyll Island Airport     31.1  -81.4    11    -5 A     America/New_Yo…
#>  6 0A9   Elizabethton Municipal …  36.4  -82.2  1593    -5 A     America/New_Yo…
#>  7 0G6   Williams County Airport   41.5  -84.5   730    -5 A     America/New_Yo…
#>  8 0G7   Finger Lakes Regional A…  42.9  -76.8   492    -5 A     America/New_Yo…
#>  9 0P2   Shoestring Aviation Air…  39.8  -76.6  1000    -5 U     America/New_Yo…
#> 10 0S9   Jefferson County Intl     48.1 -123.    108    -8 A     America/Los_An…
#> # … with 1,448 more rows

dm_nycflights13()$airports

#> # A tibble: 1,458 x 8
#>    faa   name                       lat    lon   alt    tz dst   tzone          
#>    <chr> <chr>                    <dbl>  <dbl> <dbl> <dbl> <chr> <chr>          
#>  1 04G   Lansdowne Airport         41.1  -80.6  1044    -5 A     America/New_Yo…
#>  2 06A   Moton Field Municipal A…  32.5  -85.7   264    -6 A     America/Chicago
#>  3 06C   Schaumburg Regional       42.0  -88.1   801    -6 A     America/Chicago
#>  4 06N   Randall Airport           41.4  -74.4   523    -5 A     America/New_Yo…
#>  5 09J   Jekyll Island Airport     31.1  -81.4    11    -5 A     America/New_Yo…
#>  6 0A9   Elizabethton Municipal …  36.4  -82.2  1593    -5 A     America/New_Yo…
#>  7 0G6   Williams County Airport   41.5  -84.5   730    -5 A     America/New_Yo…
#>  8 0G7   Finger Lakes Regional A…  42.9  -76.8   492    -5 A     America/New_Yo…
#>  9 0P2   Shoestring Aviation Air…  39.8  -76.6  1000    -5 U     America/New_Yo…
#> 10 0S9   Jefferson County Intl     48.1 -123.    108    -8 A     America/Los_An…
#> # … with 1,448 more rows