Load a dm from a remote data source

dm_from_src() creates a dm from some or all tables in a src (a database or an environment) or which are accessible via a DBI-Connection. For Postgres and SQL Server databases, primary and foreign keys are imported from the database.

dm_from_src(src = NULL, table_names = NULL, learn_keys = NULL, ...)

Arguments

src
A dplyr table source object or a DBI::DBIConnection object is accepted.

table_names
A character vector of the names of the tables to include.

learn_keys

Set to TRUE to query the definition of primary and foreign keys from the database. Currently works only for Postgres and SQL Server databases. The default attempts to query and issues an informative message.

...

Additional parameters for the schema learning query. Currently supports schema (default: "public") and table_type (default: "BASE TABLE") for Postgres databases.

Value

A dm object.

Examples

dm_from_src_demo <- function() {
  if (!rlang::is_installed("DBI")) {
    message("Install the DBI package to run this example.")
  }

  con <- DBI::dbConnect(
    RMariaDB::MariaDB(),
    username = "guest",
    password = "relational",
    dbname = "Financial_ijs",
    host = "relational.fit.cvut.cz"
  )
  on.exit(DBI::dbDisconnect(con))

  dm_from_src(con)
}
if (FALSE) {
dm_from_src_demo()
}