Creates a dm object for the pixarfilms data

Creates an example dm object from the tables in pixarfilms, along with the references.

dm_pixarfilms(color = TRUE)

Arguments

color

Boolean, if TRUE (default), the resulting dm object will have colors assigned to different tables for visualization with dm_draw().

Value

A dm object consisting of pixarfilms tables, complete with primary and foreign keys and optionally colored.

Examples

dm_pixarfilms()
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `pixar_films`, `pixar_people`, `academy`, `box_office`, `genres`, `public_response`
#> Columns: 23
#> Primary keys: 6
#> Foreign keys: 5
dm_pixarfilms() %>%
  dm_draw()