JSON nest join

A wrapper around dplyr::nest_join() which stores the joined data into a JSON column. json_nest_join() returns all rows and columns in x with a new JSON columns that contains all nested matches from y.

json_nest_join(x, y, by = NULL, ..., copy = FALSE, keep = FALSE, name = NULL)

Arguments

x, y

A pair of data frames or data frame extensions (e.g. a tibble).

by

A character vector of variables to join by.

If NULL, the default, *_join() will perform a natural join, using all variables in common across x and y. A message lists the variables so that you can check they're correct; suppress the message by supplying by explicitly.

To join by different variables on x and y, use a named vector. For example, by = c("a" = "b") will match x$a to y$b.

To join by multiple variables, use a vector with length > 1. For example, by = c("a", "b") will match x$a to y$a and x$b to y$b. Use a named vector to match different variables in x and y. For example, by = c("a" = "b", "c" = "d") will match x$a to y$b and x$c to y$d.

To perform a cross-join, generating all combinations of x and y, use by = character().

...

Other parameters passed onto methods.

copy

If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

keep

Should the join keys from both x and y be preserved in the output?

name

The name of the list column nesting joins create. If NULL the name of y is used.

See also

Examples

df1 <- tibble::tibble(x = 1:3)
df2 <- tibble::tibble(x = c(1, 1, 2), y = c("first", "second", "third"))
df3 <- json_nest_join(df1, df2)
#> Joining, by = "x"
df3
#> # A tibble: 3 × 2
#>       x df2       
#>   <dbl> <list>    
#> 1     1 <json [1]>
#> 2     2 <json [1]>
#> 3     3 <json [1]>
df3$df2
#> [[1]]
#> [{"y":"first"},{"y":"second"}] 
#> 
#> [[2]]
#> [{"y":"third"}] 
#> 
#> [[3]]
#> [] 
#>